[rebase] Fix API compatibility breaks from upstream uplift
- Drop const from L1SpillManagement handleNoFit/handleFragmentation/ensureFitsL1
  to resolve const-qualifier cascade in the eviction call chain
- Add (long long)/(size_t) casts on printf format args in timing instrumentation
  across analysis and optimizer pass files
- Remove dropped memory_config param from PermuteOp, ReshapeOp, ToMemoryConfigOp,
  and ToLayoutOp call sites
- Replace getMemoryConfig()|getMemoryConfig(result) with getMemoryConfigAttr()
  in GridSampleOp and PixelUnshuffleOp emit paths (TTNNToEmitC/TTNNToEmitPy)
- Remove CreateGlobalSemaphoreOp and ResetGlobalSemaphoreOp OpModel stubs
  (ops removed upstream); remove checkDeviceWorkerGrid guard from GridSampleOp
- Replace ::ttnn::TensorSpec with ::tt::tt_metal::TensorSpec and drop deviceGrid
  param from getOpConstraints (type and API moved in upstream tt_metal)
- Remove parentContext delegation from runtime implicit semaphore/grid accessors
- Fix GridSampleOp/PixelUnshuffleOp output ref to use tensorRefs vector API

+++ lib/Dialect/TTNN/Analysis/LegalOpConfigAnalysis.cpp
@@ -399,7 +399,7 @@ void LegalOpConfigAnalysis::fillOpSpecificAttrs() {
                 "[cfg-timing] Conv2d @%s done  %lld ms"
                 "  configs_generated=%zu  total_op_configs=%zu\n",
                 _locStr.c_str(),
-                std::chrono::duration_cast<std::chrono::milliseconds>(
+                (long long)std::chrono::duration_cast<std::chrono::milliseconds>(
                     std::chrono::steady_clock::now() - _tCfg)
                     .count(),
                 _cfgGenCount, analysisResult.size());
