+++ lib/Dialect/TTNN/Transforms/OptimizerPasses/TTNNPrepareConv2dWeightsAndBias.cpp
@@ -64,7 +64,7 @@ public:
                     ttnn::PrepareConv2dBiasOp>(convOp, moduleOp, rewriter);
       fprintf(stderr, "[prep2d-timing]   conv2d[%zu] done  %lld ms\n",
               _conv2dIdx,
-              std::chrono::duration_cast<std::chrono::milliseconds>(
+              (long long)std::chrono::duration_cast<std::chrono::milliseconds>(
                   std::chrono::steady_clock::now() - _t)
                   .count());
       ++_conv2dIdx;
@@ -87,7 +87,7 @@ public:
       fprintf(stderr,
               "[prep2d-timing]   conv_transpose2d[%zu] done  %lld ms\n",
               _convT2dIdx,
-              std::chrono::duration_cast<std::chrono::milliseconds>(
+              (long long)std::chrono::duration_cast<std::chrono::milliseconds>(
                   std::chrono::steady_clock::now() - _t)
                   .count());
       ++_convT2dIdx;
@@ -96,7 +96,7 @@ public:
     fprintf(stderr,
             "[prep2d-timing] TTNNPrepareConv2dWeightsAndBias TOTAL  %lld ms"
             "  conv2d=%zu  conv_transpose2d=%zu\n",
-            std::chrono::duration_cast<std::chrono::milliseconds>(
+            (long long)std::chrono::duration_cast<std::chrono::milliseconds>(
                 std::chrono::steady_clock::now() - _tTotal)
                 .count(),
             _conv2dIdx, _convT2dIdx);
